feat: Refactor image upload with GridFS and remove sharp
This commit refactors the image upload functionality to use GridFS for storage, which offers better scalability and management of large files. As part of this change, the sharp library and associated image processing logic have been removed. The multer version has been updated to 1.4.4, and multer-gridfs-storage has been added as a dependency. A middleware for processing images and directly uploading to GridFS is implemented.