Refactor: Replace `temUser` middleware with `UserMiddlewares.useUser()`
This commit removes the `temUser` middleware and replaces it with `UserMiddlewares.useUser()` in the auth and OTP routes. This aims to consolidate user retrieval logic and improve code organization.  The logic in `AuthServices.login` has also been updated to accommodate the change in request object. Also updated incorrect error messages in login service.