Refactor: Move middleware utils and improve capture middleware
This commit moves the `catchAsync` middleware to the `src/app/middlewares` directory and updates imports accordingly.  It also refactors the `capture` middleware to support parsing JSON data passed in the `data` field of the request body.  Finally, it replaces the file deletion utilities with a `deleteImage` function located in the capture middleware, improving error handling during file removal in the global error handler and user service. The Redis client configuration was also commented out.